opengl/requirements.txt
cx-freeze
esky
PySide
smartside
pyopengl
euclid3
